requirements-dev.txt
types-Pillow
pillow